backend/requirements.txt
# FastAPI 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 交易所 API
ccxt>=4.2.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
aiosqlite>=0.19.0

# 缓存 (可选，后续添加)
# redis>=5.0.0

# 定时任务
apscheduler>=3.10.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 加密
cryptography>=41.0.0

# 日志
loguru>=0.7.0

# HTTP 客户端
httpx>=0.25.0
aiohttp>=3.9.0

# WebSocket
websockets>=12.0

# AI (可选)
# dashscope>=1.14.0
